+++ src/sidebar/activateMod/ActivateModInputs.tsx
