+++ src/sidebar/activateMod/ActivateModPanel.module.scss
