+++ src/sidebar/activateMod/ActivateMultipleModsPanel.test.tsx
