+++ src/sidebar/activateMod/RequireMods.tsx
